C0006_xinference/requirements.txt
xinference-client==1.5.0